\name{bronchitis}
\docType{data}
\alias{bronchit}
\title{Chronic bronchitis in a sample of mean in Cardiff}
\description{
  The data consist of observations on three variables
for each of 212 men in a sample of Cardiff enumeration districts.
}
\usage{data(bronchit)}
\format{A data.frame of 212 obs of 3 variables:
  \tabular{rll}{
    [,1] \tab r \tab integer, 1= respondent suffered from chronic bronchitis \cr
    [,2] \tab cig \tab numeric, the number of cigarettes per day\cr
    [,3] \tab poll \tab numeric, the smoke level in the locality
    }
  }
\source{Jones, K. (1975), \emph{A geographical contribution to the aetiology of
                  chronic bronchitis}, Unpublished BSc dissertation,
		University of Southampton. Published in
		Wrigley, N. (1976). \emph{Introduction to the use of
		  logit models in geography}, Geo.Abstracts Ltd, CATMOG
		10, University of East Anglia, Norwich.
	      }
\note{See p.224 in SMIR}
\keyword{datasets}