\name{DAAGviz-package}
\alias{DAAGviz-package}
\alias{DAAGviz}
\docType{package}
\title{Functions, data and vignettes for a course on R visualization
}
\description{
The main purpose is to provide functions that can be used to reproduce all graphs in the course notes.
}
\details{
\tabular{ll}{
Package: \tab DAAGviz\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-12-19\cr
License: \tab GPL (>=2)\cr
}
Use mainly to provide functions that can be used to reproduce all graphs in the course notes.
}
\author{
John Maindonald

Maintainer: John Maindonald <john.maindonald@anu.edu.au>
}
\keyword{ package }

